Refund payment

See merge request kkm_android/ofdapp!44

+++ data/src/main/java/kg/nurtelecom/data/history/Content.kt
@@ -6,7 +6,7 @@ data class Content (
 	val total : BigDecimal,
 	val gnsRegNum : Int,
 	val indexNum : Int,
-	val subtotal : Int,
+	val subtotal : Double,
 	val operationType : String,
 	val totalGoodsSum : Double,
 	val createdAt : String,
